+++ df-downloader-ui/src/routes/settings/settings.component.tsx
@@ -1,10 +1,11 @@
-import { Box, useMediaQuery } from "@mui/material";
-import { theme } from "../../themes/theme";
+import { Box, useMediaQuery,
+  useTheme } from "@mui/material";
 
 export type SettingsElementProps = {
   children: React.ReactNode;
 };
 export const SettingsElement = ({ children }: SettingsElementProps) => {
+  const theme = useTheme();
   const useMobileLayout = useMediaQuery(theme.breakpoints.down("lg"));
   return <Box sx={{ display: "flex", width: useMobileLayout ? "100%" : theme.breakpoints.values.md }}>{children}</Box>;
 };